+++ Models/Productos.cs
@@ -5,11 +5,11 @@ namespace tl2_tp8_2025_BautistaAlvarez.Models
         private int idProducto;
 
         private string descripcion;
-        private int precio;
+        private decimal precio;
         public int IdProducto { get => idProducto; set => idProducto = value; }
         public string Descripcion { get => descripcion; set => descripcion = value; }
 
-        public int Precio { get => precio; set => precio = value; }
+        public decimal Precio { get => precio; set => precio = value; }
 
         public Productos()
         {
